delete proxy module in preBundle
For the old version of "esbuild", in order to flatten the dependent pre built products, the agent module will be used to process the "require ('react /jsx runtime. js')". The reason why the path was not directly modified to redirect to the entry is that the products will have a nested structure. The subsequent packaging is also added to prevent secondary packaging.
However, for the current version of "esbuild", even if it is directly redirected to the entry module, the product name is still specified by entryPoints and will not result in secondary packaging or nested products. Therefore, we do not need to use the proxy module now.
After the above code is deleted and added, vite still runs well, and the pre built products and product size will not change. Even if module "A" depends on module "B", using modules A and B in the project can also extract the same part.
Deleting the agent module will optimize more volumes.